# perf: speedup `@redis_cache` by 2x-20x