We make 2 redis call instead of 1. This was done to ensure functions returning `None` can be cached. 

BUT this means everything now has this overhead of almost 2x redis cost. This PR rearranges the check in a way that most common reads will now not have to pay this extra cost.

Microbenchmarks results:

First access is faster by 2x.
bench_utils_bench_redis_cache_deco_without_local_cache: 13.0 ms +- 0.1 ms -> 6.60 ms +- 0.06 ms: 1.98x faster

Repeat accesses are faster by 24x :woozy_face: 
bench_utils_bench_redis_cache_deco_with_local_cache: 6.76 ms +- 0.07 ms ->  281 us +- 2 us: 24.07x faster


Closes https://github.com/frappe/caffeine/issues/11 


Unrelated potentially _breaking_ fixes:

- Expirable keys are always cached in `frappe.local`. This check is also nonsensical since they get cached on 2nd run anyway. Think of this as "repeatable read" if you want to. :woozy_face: